GSYE-880: Move heatpump tank state modules into separate package
